separate functions and create localStorageUserId hook
